refactor(gas): move Somnia estimation driver to its own module
The one-probe-per-eth_call driver, its bisection helper, and
SOMNIA_CHAIN_IDS move from the chain-agnostic GasManager base class to
voltaire_bundler/gas/somnia_gas_estimation.py, with the Somnia gas
model documented in the module docstring. The driver becomes a free
function taking the gas manager instance, which also removes the
simulate_handle_op_mod stub the base class only carried for the
driver's benefit. No behavior change; tests re-point their
eth_blockNumber mock to the new module.
